yodack - the beginning
This is the first import of the only *Kit that does not suck: Ye Olde
Debian Archive Control Kit.

The tool works, mostly, there is even some documentation (but the
configuration needs to be documented better), but a lot of things need
improvement still.

From error handling, through messages, to documentation.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>